+++ tests/integration/plugins/test_python.py
@@ -35,6 +35,31 @@ def teardown_module():
     plugins.unregister_all()
 
 
+def test_python_plugin(new_dir):
+    """Prime a simple python source."""
+    source_location = Path(__file__).parent / "test_python"
+
+    parts_yaml = textwrap.dedent(
+        f"""\
+        parts:
+          foo:
+            plugin: python
+            source: {source_location}
+        """
+    )
+    parts = yaml.safe_load(parts_yaml)
+
+    lf = LifecycleManager(parts, application_name="test_python", cache_dir=new_dir)
+    actions = lf.plan(Step.PRIME)
+
+    with lf.action_executor() as ctx:
+        ctx.execute(actions)
+
+    primed_script = Path(lf.project_info.prime_dir, "bin", "mytest")
+    assert primed_script.exists()
+    assert primed_script.open().readline().rstrip() == "#!/usr/bin/env python3"
+
+
 def test_python_plugin_symlink(new_dir):
     """Run in the standard scenario with no overrides."""
     parts_yaml = textwrap.dedent(